some quality of life updates
