+++ src/payload-config/i18n/locales/en.json
@@ -264,6 +264,7 @@
     "confirmedSuccessfully": "Confirmed successfully",
     "zone": "Zone:",
     "ticketCode": "Ticket Code:",
+    "seat": "Seat:",
     "eventName": "Event Name:",
     "email": "Email:",
     "attendeeName": "Attendee Name:",
