+++ src/payload-config/i18n/locales/vi.json
@@ -264,6 +264,7 @@
     "confirmedSuccessfully": "Đã xác nhận thành công!",
     "zone": "Khu vực:",
     "ticketCode": "Mã vé:",
+    "seat": "Ghế:",
     "eventName": "Tên sự kiện:",
     "email": "Email:",
     "attendeeName": "Tên người tham dự:",
